refactor: extract commandTreeUtils
