+++ src/web/src/views/cli/GenerateDialog.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Alert, Box, Button, Dialog, DialogActions, DialogContent, DialogTitle, LinearProgress } from "@mui/material";
 import { cliApi, errorHandlerApi } from "../../services";
-import { exportModViewProfile, type ProfileCommandTree } from "./CLIModGeneratorProfileCommandTree";
+import { exportModViewProfile, type ProfileCommandTree } from "./utils/commandTreeInitialization";
 import { type CLIModViewProfiles } from "./interfaces";
 
 interface ProfileCommandTrees {
